requirements.txt
websocket-client
numpy
python-binance
pandas